checkpoint for code mirror
