+++ Gopkg.lock
@@ -456,6 +456,6 @@
 [solve-meta]
   analyzer-name = "dep"
   analyzer-version = 1
-  inputs-digest = "bfd6a8ed68b1e61333e0ea442d328f0f4511ebd7c698e1685ab417423be59bb7"
+  inputs-digest = "606d07a5d2a13f3f6b62cb6670848ab1628a141d5a05b4e7e453898606447605"
   solver-name = "gps-cdcl"
   solver-version = 1
